Move deck import hook into model
