+++ src/features/deck-import/model/useDeckImport.spec.tsx
@@ -5,7 +5,7 @@ import { act, renderHook, waitFor } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { createCard, createDeck, createLocalCard, createLocalDeck } from "@/test/factories";
-import type { DeckImportResult } from "../model/deckImportExecution";
+import type { DeckImportResult } from "./deckImportExecution";
 import { actAsync } from "@/test/act";
 
 const mocks = vi.hoisted(() => ({
